+++ lib/src/models/index.dart
@@ -1,3 +1,4 @@
+export 'audio_track.dart';
 export 'option_item.dart';
 export 'options_translation.dart';
 export 'subtitle_model.dart';